Clean up the FilterBar implementation
JIRA: RAIL-3382
